ln-simln-jamming/tests: move reputation interceptor mock to utils
